backend/pyproject.toml
[tool.poetry]
name = "g-guid"
version = "0.1.0"
description = ""
authors = ["gvensye <bnlvdn7483@gmail.com>"]
readme = "readme.md"
packages = [{include = "g_guid"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.106.0"
uvicorn = "^0.23.2"
sqlalchemy = "2.0.13"
psycopg2-binary = "2.9.6"
python-dotenv = "^1.0.0"
alembic = "1.11.1"
passlib = "^1.7.4"
python-slugify = "^8.0.1"
pytest = "^7.4.2"
httpx = "^0.25.0"
pydantic-settings = "^2.0.3"
pre-commit = "^3.4.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
fastapi-debug-toolbar = "^0.6.2"


[tool.poetry.group.dev.dependencies]
reorder-python-imports = "^3.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
